+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/ContentManagerPlugin.java
@@ -63,6 +63,7 @@ import java.util.List;
 import java.util.function.Supplier;
 
 import com.wazuh.contentmanager.cti.catalog.index.ConsumersIndex;
+import com.wazuh.contentmanager.cti.catalog.service.LogtestService;
 import com.wazuh.contentmanager.cti.catalog.service.SecurityAnalyticsService;
 import com.wazuh.contentmanager.cti.catalog.service.SecurityAnalyticsServiceImpl;
 import com.wazuh.contentmanager.cti.catalog.service.SpaceService;
@@ -98,6 +99,7 @@ public class ContentManagerPlugin extends Plugin
     private EngineService engine;
     private SpaceService spaceService;
     private SecurityAnalyticsService securityAnalyticsService;
+    private LogtestService logtestService;
 
     /**
      * Initializes the plugin components, including the CTI console, consumer index helpers, and the
@@ -167,6 +169,9 @@ public class ContentManagerPlugin extends Plugin
             this.securityAnalyticsService = new SecurityAnalyticsServiceImpl(client);
         }
 
+        this.logtestService =
+                new LogtestService(this.engine, this.securityAnalyticsService, this.client);
+
         // Register hot-reload settings consumer
         clusterService
                 .getClusterSettings()
@@ -233,8 +238,8 @@ public class ContentManagerPlugin extends Plugin
                 new RestPostSubscriptionAction(this.ctiConsole),
                 new RestDeleteSubscriptionAction(this.ctiConsole),
                 new RestPostUpdateAction(this.ctiConsole, this.catalogSyncJob),
-                // User-generated content endpoints (Logtest)
-                new RestPostLogtestAction(this.engine),
+                // User-generated content endpoints
+                new RestPostLogtestAction(this.logtestService),
                 // Policy endpoints
                 new RestPutPolicyAction(this.spaceService, this.engine),
                 // Rule endpoints